Refactor classes to remove EDMReceiver.java
