+++ src/main/java/de/wpsverlinden/c4cduplicateanalyzer/feed/ODataFeedReceiver.java
@@ -4,24 +4,27 @@ import de.wpsverlinden.c4cduplicateanalyzer.ApplicationConfiguration;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Optional;
+import javax.annotation.PostConstruct;
 import org.apache.olingo.odata2.api.edm.Edm;
 import org.apache.olingo.odata2.api.edm.EdmEntityContainer;
 import org.apache.olingo.odata2.api.ep.EntityProvider;
+import org.apache.olingo.odata2.api.ep.EntityProviderException;
 import org.apache.olingo.odata2.api.ep.EntityProviderReadProperties;
 import org.apache.olingo.odata2.api.ep.entry.ODataEntry;
 import org.apache.olingo.odata2.api.ep.feed.ODataFeed;
 import org.apache.olingo.odata2.api.exception.ODataException;
 import org.slf4j.Logger;
+import org.springframework.batch.core.StepContribution;
+import org.springframework.batch.core.scope.context.ChunkContext;
+import org.springframework.batch.core.step.tasklet.Tasklet;
+import org.springframework.batch.repeat.RepeatStatus;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 @Component
-public class ODataFeedReceiver {
+public class ODataFeedReceiver implements Tasklet {
 
     public static final String SEPARATOR = "/";
-
-    @Autowired
-    private EDMReceiver edmReceiver;
 
     @Autowired
     private HttpConnectionHelper httpConnectionHelper;
@@ -32,10 +35,20 @@ public class ODataFeedReceiver {
     @Autowired
     private ApplicationConfiguration config;
 
+    private Edm edm;
+
+    @PostConstruct
+    private void initialize() throws IOException, EntityProviderException {
+        final String metadata_endpoint = config.getEndpoint() + "/$metadata";
+        LOG.debug("Requesting service metadata from '{}'", metadata_endpoint);
+        InputStream content = httpConnectionHelper.getConnectionInputStream(metadata_endpoint, HttpConnectionHelper.APPLICATION_XML, config.getUser(), config.getPassword());
+        this.edm = EntityProvider.readMetadata(content, true);
+        LOG.debug("Received service metadata '{}'", edm);
+    }
+
     public ODataFeed readFeed(String entitySetName, Optional<String> urlParameter) throws IOException, ODataException {
-        Edm edm = edmReceiver.getEdm();
         EdmEntityContainer entityContainer = edm.getDefaultEntityContainer();
-        String targetUri = buildUri(edmReceiver.getEndpoint(), entitySetName, urlParameter);
+        String targetUri = buildUri(config.getEndpoint(), entitySetName, urlParameter);
         LOG.debug("Generated request url '{}'", targetUri);
         InputStream content = httpConnectionHelper.getContentInputStream(targetUri, HttpConnectionHelper.APPLICATION_JSON, config.getUser(), config.getPassword());
         return EntityProvider.readFeed(HttpConnectionHelper.APPLICATION_JSON,
@@ -45,9 +58,8 @@ public class ODataFeedReceiver {
     }
 
     public ODataEntry readEntry(String entitySetName, Optional<String> urlParameter) throws IOException, ODataException {
-        Edm edm = edmReceiver.getEdm();
         EdmEntityContainer entityContainer = edm.getDefaultEntityContainer();
-        String targetUri = buildUri(edmReceiver.getEndpoint(), entitySetName, urlParameter);
+        String targetUri = buildUri(config.getEndpoint(), entitySetName, urlParameter);
         LOG.debug("Generated request url '{}'", targetUri);
         InputStream content = httpConnectionHelper.getContentInputStream(targetUri, HttpConnectionHelper.APPLICATION_JSON, config.getUser(), config.getPassword());
         return EntityProvider.readEntry(HttpConnectionHelper.APPLICATION_JSON,
@@ -62,4 +74,14 @@ public class ODataFeedReceiver {
         return targetUri.toString();
     }
 
+    @Override
+    public RepeatStatus execute(StepContribution contribution, ChunkContext chunkContext) throws Exception {
+        try {
+            initialize();
+        } catch (IOException | EntityProviderException e) {
+            LOG.error("Error initializing EDM: {}", e);
+            throw e;
+        }
+        return RepeatStatus.FINISHED;
+    }
 }
